Integrate generation with SQF and refactor
