+++ central/reports/scheduler/v2/reportgenerator/types.go
@@ -26,18 +26,25 @@ type reportEmailSubjectFormat struct {
 }
 
 type ImageCVEQueryResponse struct {
-	Cluster        string                        `db:"cluster"`
-	Namespace      string                        `db:"namespace"`
-	Deployment     string                        `db:"deployment"`
-	Image          string                        `db:"image"`
-	Component      string                        `db:"component"`
-	CVEID          string                        `db:"cve_id"`
-	CVE            string                        `db:"cve"`
-	Fixable        bool                          `db:"fixable"`
-	FixedByVersion string                        `db:"fixed_by"`
-	Severity       storage.VulnerabilitySeverity `db:"severity"`
-	CVSS           float64                       `db:"cvss"`
-	DiscoveredAt   *time.Time                    `db:"first_image_occurrence_timestamp"`
+	Cluster           string                        `db:"cluster"`
+	Namespace         string                        `db:"namespace"`
+	Deployment        string                        `db:"deployment"`
+	Image             string                        `db:"image"`
+	Component         string                        `db:"component"`
+	ComponentVersion  string                        `db:"component_version"`
+	CVEID             string                        `db:"cve_id"`
+	CVE               string                        `db:"cve"`
+	Fixable           bool                          `db:"fixable"`
+	FixedByVersion    string                        `db:"fixed_by"`
+	Severity          storage.VulnerabilitySeverity `db:"severity"`
+	CVSS              float64                       `db:"cvss"`
+	DiscoveredAtImage *time.Time                    `db:"first_image_occurrence_timestamp"`
 
 	Link string
+}
+
+type ReportData struct {
+	CVEResponses            []*ImageCVEQueryResponse
+	NumDeployedImageResults int
+	NumWatchedImageResults  int
 }
